+++ ldfparser/lin.py
@@ -3,6 +3,7 @@ Utility classes for LIN objects
 """
 from typing import Union
 
+
 class LinVersion:
     """
     LinVersion represents the LIN protocol and LDF language versions
@@ -40,6 +41,8 @@ class LinVersion:
     def __eq__(self, o: object) -> bool:
         if isinstance(o, LinVersion):
             return self.major == o.major and self.minor == o.minor
+        elif isinstance(o, J2602Version):
+            return self.major == 2 and self.minor == 0
         return False
 
     def __gt__(self, o) -> bool:
@@ -49,6 +52,8 @@ class LinVersion:
             if self.major == o.major and self.minor > o.minor:
                 return True
             return False
+        elif isinstance(o, J2602Version):
+            return (self.major == 2 and self.minor > 0) or self.major > 2
         raise TypeError()
 
     def __lt__(self, o) -> bool:
@@ -58,6 +63,8 @@ class LinVersion:
             if self.major == o.major and self.minor < o.minor:
                 return True
             return False
+        elif isinstance(o, J2602Version):
+            return self.major < 2
         raise TypeError()
 
     def __ge__(self, o) -> bool:
@@ -118,11 +125,103 @@ class Iso17987Version:
 
 ISO17987_2015 = Iso17987Version(2015)
 
-def parse_lin_version(version: str) -> Union[LinVersion, Iso17987Version]:
-    try:
-        return LinVersion.from_string(version)
-    except ValueError:
+class J2602Version:
+    def __init__(self, major, minor, part):
+        """
+        Abstract the J2602 version.
+        """
+        self.major = major
+        self.minor = minor
+        self.part = part
+
+    @staticmethod
+    def from_string(version: str) -> 'J2602Version':
+        """
+        Create an instance from the version string.
+
+        The version string J2602_1_2.0 will render:
+            major=2, minor=0, part=1
+
+        Support for J2602_1_2.0 is not implemented at this time.
+        """
+        if "J2602" not in version:
+            raise ValueError(f'{version} is not an SAE J2602 version.')
+
+        version = version.replace("J2602_", '')
+        (part, versions) = version.split('_')
+        major, minor = [int(value) for value in versions.split('.')]
+        if major == 1:
+            return J2602Version(major=major, minor=minor, part=int(part))
+
+        raise ValueError(f'{version} is not supported yet.')
+
+    def __str__(self) -> str:
+        return f"J2602_{self.part}_{self.major}.{self.minor}"
+
+    def __eq__(self, o: object) -> bool:
+        """
+        According to J2602-3_202110, section 7.1.3:
+        “J2602_1_1.0” -> J2602:2012 and earlier -> based on LIN 2.0
+        “J2602_1_2.0” -> J2602:2021 -> based on ISO 17987:2016
+
+        Therefore,
+        “J2602_1_1.0” is considered equal to LinVersion(2, 0)
+
+        """
+        if isinstance(o, J2602Version):
+            return (
+                self.major == o.major and
+                self.minor == o.minor and
+                self.part == o.part
+            )
+        elif isinstance(o, Iso17987Version):
+            return False
+        elif isinstance(o, LinVersion):
+            return o == LIN_VERSION_2_0
+        return False
+
+    def __gt__(self, o) -> bool:
+        if isinstance(o, J2602Version):
+            return (
+                self.major > o.major or
+                (
+                    self.major == o.major and self.minor > o.minor
+                )
+            )
+        if isinstance(o, Iso17987Version):
+            return False
+        if isinstance(o, LinVersion):
+            return o < LIN_VERSION_2_0
+        raise TypeError()
+
+    def __lt__(self, o) -> bool:
+        if isinstance(o, J2602Version):
+            return (
+                self.major < o.major or
+                (
+                    self.major == o.major and self.minor < o.minor
+                )
+            )
+        if isinstance(o, Iso17987Version):
+            return True
+        if isinstance(o, LinVersion):
+            return o > LIN_VERSION_2_0
+        raise TypeError()
+
+    def __ge__(self, o) -> bool:
+        return not self.__lt__(o)
+
+    def __le__(self, o) -> bool:
+        return not self.__gt__(o)
+
+    def __ne__(self, o: object) -> bool:
+        return not self.__eq__(o)
+
+def parse_lin_version(version: str) -> Union[LinVersion, Iso17987Version, J2602Version]:
+    for version_class in [LinVersion, Iso17987Version, J2602Version]:
         try:
-            return Iso17987Version.from_string(version)
-        except ValueError:
-            raise ValueError(f'{version} is not a valid LIN version.')
+            return version_class.from_string(version)
+        except (ValueError, IndexError):
+            pass
+
+    raise ValueError(f'{version} is not a valid LIN version.')